requirements.txt
altgraph==0.17
future==0.18.2
lxml==4.5.1
mss==5.1.0
pefile==2019.4.18
Pillow==7.2.0
PyInstaller==3.6
python-docx==0.8.10
pywin32-ctypes==0.2.0
selenium==3.141.0
urllib3==1.25.9
